Use ResponseEntity

+++ src/main/java/com/gatherProject/GatherBackend/Controllers/EventController.java
@@ -3,12 +3,14 @@ package com.gatherProject.GatherBackend.Controllers;
 import com.gatherProject.GatherBackend.Models.Event;
 import com.gatherProject.GatherBackend.Services.EventService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.concurrent.ExecutionException;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/events")
 public class EventController {
     final EventService eventService;
 
@@ -17,23 +19,37 @@ public class EventController {
         this.eventService = eventService;
     }
 
-    @PostMapping("/events")
-    public Event createEvent(@RequestBody Event event) throws ExecutionException, InterruptedException {
-        return eventService.createEvent(event);
+    @PostMapping
+    public ResponseEntity<Event> createEvent(@RequestBody Event event) throws ExecutionException, InterruptedException {
+        return ResponseEntity.ok(eventService.createEvent(event));
     }
 
-    @GetMapping("/events/{id}")
-    public Event readEvent(@PathVariable String id) throws ExecutionException, InterruptedException {
-        return eventService.getEvent(id);
+    @GetMapping("/{id}")
+    public ResponseEntity<Event> readEvent(@PathVariable String id) throws ExecutionException, InterruptedException {
+        Event event = eventService.getEvent(id);
+
+        if (event == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
+        return ResponseEntity.ok(event);
     }
 
-    @PutMapping("/events")
-    public Event updateEvent(@RequestBody Event event) throws ExecutionException, InterruptedException {
-        return eventService.updateEvent(event);
+    @PutMapping
+    public ResponseEntity<Event> updateEvent(@RequestBody Event event) throws ExecutionException, InterruptedException {
+        return ResponseEntity.ok(eventService.updateEvent(event));
     }
 
-    @DeleteMapping("/events/{id}")
-    public void deleteEvent(@PathVariable String id) throws ExecutionException, InterruptedException {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Event> deleteEvent(@PathVariable String id) throws ExecutionException, InterruptedException {
+        Event event = eventService.getEvent(id);
+
+        if (event == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
         eventService.deleteEvent(id);
+
+        return ResponseEntity.ok(event);
     }
 }
